+++ spec/acceptance/se/repository_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'spec_helper_acceptance'
 
 describe 'graphdb::se::repository', unless: UNSUPPORTED_PLATFORMS.include?(fact('osfamily')) do