+++ spec/classes/graphdb_spec.rb
@@ -1,14 +1,16 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe 'graphdb', type: :class do
   on_supported_os.each do |os, facts|
     context "on #{os}" do
       let(:facts) do
-        facts.merge(machine_java_home: '/opt/jdk7')
+        facts.merge(machine_java_home: '/opt/jdk8')
       end
 
       describe 'with minimum configuration' do
-        let(:params) { { version: '7.0.0', edition: 'ee' } }
+        let(:params) { { version: '9.10.1', edition: 'ee' } }
 
         it { is_expected.to contain_class('graphdb::params') }
         it {   is_expected.to contain_package('unzip') }
@@ -30,7 +32,7 @@ describe 'graphdb', type: :class do
       end
 
       describe 'with custom graphdb_user and graphdb_group' do
-        let(:params) { { version: '7.0.0', edition: 'ee', graphdb_user: 'user', graphdb_group: 'group' } }
+        let(:params) { { version: '9.10.1', edition: 'ee', graphdb_user: 'user', graphdb_group: 'group' } }
 
         it { is_expected.to contain_file('/opt/graphdb').with(ensure: 'directory', owner: 'user', group: 'group') }
         it { is_expected.to contain_file('/var/tmp/graphdb').with(ensure: 'directory', owner: 'user', group: 'group') }
@@ -55,7 +57,7 @@ describe 'graphdb', type: :class do
         end
 
         context 'unknown ensure' do
-          let(:params) { { version: '7.0.0', edition: 'ee', ensure: 'test' } }
+          let(:params) { { version: '9.10.1', edition: 'ee', ensure: 'test' } }
 
           it do
             expect { is_expected.to contain_class('graphdb') }.to raise_error(Puppet::ParseError)
@@ -63,7 +65,7 @@ describe 'graphdb', type: :class do
         end
 
         context 'unknown status' do
-          let(:params) { { version: '7.0.0', edition: 'ee', status: 'test' } }
+          let(:params) { { version: '9.10.1', edition: 'ee', status: 'test' } }
 
           it do
             expect { is_expected.to contain_class('graphdb') }.to raise_error(Puppet::ParseError)
@@ -71,7 +73,7 @@ describe 'graphdb', type: :class do
         end
 
         context 'not absolute data_dir path' do
-          let(:params) { { version: '7.0.0', edition: 'ee', data_dir: 'test' } }
+          let(:params) { { version: '9.10.1', edition: 'ee', data_dir: 'test' } }
 
           it do
             expect { is_expected.to contain_class('graphdb') }.to raise_error(Puppet::ParseError)
@@ -79,7 +81,7 @@ describe 'graphdb', type: :class do
         end
 
         context 'not absolute tmp_dir path' do
-          let(:params) { { version: '7.0.0', edition: 'ee', tmp_dir: 'test' } }
+          let(:params) { { version: '9.10.1', edition: 'ee', tmp_dir: 'test' } }
 
           it do
             expect { is_expected.to contain_class('graphdb') }.to raise_error(Puppet::ParseError)
@@ -87,7 +89,7 @@ describe 'graphdb', type: :class do
         end
 
         context 'not absolute install_dir path' do
-          let(:params) { { version: '7.0.0', edition: 'ee', install_dir: 'test' } }
+          let(:params) { { version: '9.10.1', edition: 'ee', install_dir: 'test' } }
 
           it do
             expect { is_expected.to contain_class('graphdb') }.to raise_error(Puppet::ParseError)
@@ -95,7 +97,7 @@ describe 'graphdb', type: :class do
         end
 
         context 'not valid manage_graphdb_user' do
-          let(:params) { { version: '7.0.0', edition: 'ee', manage_graphdb_user: 'test' } }
+          let(:params) { { version: '9.10.1', edition: 'ee', manage_graphdb_user: 'test' } }
 
           it do
             expect { is_expected.to contain_class('graphdb') }.to raise_error(Puppet::ParseError)
@@ -103,7 +105,7 @@ describe 'graphdb', type: :class do
         end
 
         context 'not valid graphdb_user' do
-          let(:params) { { version: '7.0.0', edition: 'ee', graphdb_user: '' } }
+          let(:params) { { version: '9.10.1', edition: 'ee', graphdb_user: '' } }
 
           it do
             expect { is_expected.to contain_class('graphdb') }.to raise_error(Puppet::ParseError)
@@ -111,7 +113,7 @@ describe 'graphdb', type: :class do
         end
 
         context 'not valid graphdb_group' do
-          let(:params) { { version: '7.0.0', edition: 'ee', graphdb_group: '' } }
+          let(:params) { { version: '9.10.1', edition: 'ee', graphdb_group: '' } }
 
           it do
             expect { is_expected.to contain_class('graphdb') }.to raise_error(Puppet::ParseError)
@@ -119,7 +121,7 @@ describe 'graphdb', type: :class do
         end
 
         context 'not valid purge_data_dir' do
-          let(:params) { { version: '7.0.0', edition: 'ee', purge_data_dir: 'test' } }
+          let(:params) { { version: '9.10.1', edition: 'ee', purge_data_dir: 'test' } }
 
           it do
             expect { is_expected.to contain_class('graphdb') }.to raise_error(Puppet::ParseError)
@@ -127,7 +129,7 @@ describe 'graphdb', type: :class do
         end
 
         context 'not valid archive_dl_timeout' do
-          let(:params) { { version: '7.0.0', edition: 'ee', archive_dl_timeout: 'test' } }
+          let(:params) { { version: '9.10.1', edition: 'ee', archive_dl_timeout: 'test' } }
 
           it do
             expect { is_expected.to contain_class('graphdb') }.to raise_error(Puppet::ParseError)
@@ -135,7 +137,7 @@ describe 'graphdb', type: :class do
         end
 
         context 'not valid java_home' do
-          let(:params) { { version: '7.0.0', edition: 'ee', java_home: 'test' } }
+          let(:params) { { version: '9.10.1', edition: 'ee', java_home: 'test' } }
 
           it do
             expect { is_expected.to contain_class('graphdb') }.to raise_error(Puppet::ParseError)