+++ spec/defines/service_spec.rb
@@ -1,17 +1,19 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe 'graphdb::service', type: :define do
   let :default_facts do
     {
       kernel: 'Linux',
-      machine_java_home: '/opt/jdk7'
+      machine_java_home: '/opt/jdk8'
     }
   end
 
   let(:title) { 'test' }
 
   let :pre_condition do
-    "class { 'graphdb': version => '7.0.0', edition => 'ee' }"
+    "class { 'graphdb': version => '9.10.1', edition => 'ee' }"
   end
 
   { 'enabled' => ['running', true],
@@ -23,7 +25,7 @@ describe 'graphdb::service', type: :define do
       let(:facts) { default_facts.merge(operatingsystem: 'Ubuntu', operatingsystemmajrelease: '12') }
 
       context "with ensure set to: [present] and status set to: [#{status_param}]" do
-        let(:params) { { ensure: 'present', status:  status_param } }
+        let(:params) { { ensure: 'present', status: status_param } }
 
         it do
           is_expected.to contain_graphdb__service__upstart(title).with(
@@ -47,8 +49,8 @@ describe 'graphdb::service', type: :define do
       end
     end
 
-    context 'Debian 9' do
-      let(:facts) { default_facts.merge(operatingsystem: 'Debian', operatingsystemmajrelease: '9') }
+    context 'Debian 10' do
+      let(:facts) { default_facts.merge(operatingsystem: 'Debian', operatingsystemmajrelease: '10') }
 
       context "with ensure set to: [present] and status set to: [#{status_param}]" do
         let(:params) { { ensure: 'present', status:  status_param } }
@@ -75,8 +77,8 @@ describe 'graphdb::service', type: :define do
       end
     end
 
-    context 'Debian 5' do
-      let(:facts) { default_facts.merge(operatingsystem: 'Debian', operatingsystemmajrelease: '5') }
+    context 'Debian 7' do
+      let(:facts) { default_facts.merge(operatingsystem: 'Debian', operatingsystemmajrelease: '7') }
 
       context "with ensure set to: [present] and status set to: [#{status_param}]" do
         let(:params) { { ensure: 'present', status:  status_param } }