+++ spec/defines/services/systemd_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe 'graphdb::service::systemd', type: :define do
@@ -6,15 +8,15 @@ describe 'graphdb::service::systemd', type: :define do
       kernel: 'Linux',
       operatingsystem: 'Ubuntu',
       operatingsystemmajrelease: '16',
-      machine_java_home: '/opt/jdk7',
-      ipaddress: '127.0.0.1'
+      machine_java_home: '/opt/jdk8',
+      ipaddress: '129.10.1.1'
     }
   end
 
   let(:title) { 'test' }
 
   let :pre_condition do
-    "class { 'graphdb': version => '7.0.0', edition => 'ee' }"
+    "class { 'graphdb': version => '9.10.1', edition => 'ee' }"
   end
 
   context 'with ensure set to present' do
@@ -43,7 +45,7 @@ describe 'graphdb::service::systemd', type: :define do
     end
     it do
       is_expected.to contain_file('/lib/systemd/system/graphdb-test.service')
-        .with(content: /JAVA_HOME=\/opt\/jdk7/)
+        .with(content: /JAVA_HOME=\/opt\/jdk8/)
     end
     it do
       is_expected.to contain_file('/lib/systemd/system/graphdb-test.service')