+++ spec/functions/generate_java_opts_string_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 
 describe 'generate_java_opts_string', type: :puppet_function do