+++ spec/unit/util/request_manager_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require 'spec_helper'
 require 'puppet/util/request_manager'
 
@@ -62,7 +64,7 @@ describe '#perform_http_request' do
                                           .to_return(status: [404]).then
                                           .to_return(status: [501]).then
 
-      expect { call_perform_http_request }.to raise_error(Puppet::Error, /Unrecoverable response recieved/)
+      expect { call_perform_http_request }.to raise_error(Puppet::Error, /Unrecoverable response received/)
       expect(WebMock).to have_requested(method, uri).times(3)
     end
   end